Dress NPCs,Allows you to equip gear on NPCs. No client modifications!
Set the displays of items you want NPC to have and his race and gender as well as facial features and skin color. All this is done in the database.

Note that the items are displayids, not item entries. Remember to clear cache when using on an npc :3see:http://rochet2.github.io/?page=Dress_NPCs

Signed-off-by: EdwardTuring <mxturing@yeah.net>